+++ LOVE-WrapLua/OneLua/graphics.lua
@@ -17,6 +17,7 @@ lv1lua.loadOnce("LOVE-WrapLua/core/util.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/transform.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/textwrap.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/polyfill.lua")
+lv1lua.loadOnce("LOVE-WrapLua/core/capabilities.lua")
 
 local GRAPHICS = "LOVE-WrapLua/OneLua/graphics/"