+++ LOVE-WrapLua/PS3/graphics.lua
@@ -13,6 +13,7 @@ end
 
 lv1lua.loadOnce("LOVE-WrapLua/core/util.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/textwrap.lua")
+lv1lua.loadOnce("LOVE-WrapLua/core/capabilities.lua")
 
 local GRAPHICS = "LOVE-WrapLua/PS3/graphics/"