+++ LOVE-WrapLua/lpp-vita/graphics.lua
@@ -12,6 +12,7 @@ lv1lua.loadOnce("LOVE-WrapLua/core/util.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/transform.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/textwrap.lua")
 lv1lua.loadOnce("LOVE-WrapLua/core/polyfill.lua")
+lv1lua.loadOnce("LOVE-WrapLua/core/capabilities.lua")
 
 local GRAPHICS = "LOVE-WrapLua/lpp-vita/graphics/"